Refactor ability effects with helper functions
